The First iteration of the Group
By randomly generating a 55x55 grid of small squares with red, yellow, and blue as the main colors and white and gray as the secondary colors to form the base, and adding about 40% of the total area of the blank squares to form the blank area in the artwork, and finally generating about 15 large squares as the top irregular rectangles in the layer of “Broadway Boogie Woogie” as a decorative feature, but then discovered that in the center of these irregular rectangles there are also some small squares of different colors, and similarly adding these small squares of different colors into the large squares will follow the random generation of the large squares. However, it was found that there are some small squares of different colors in the center of these irregular rectangles. Similarly, adding these small squares of different colors inside the big squares will follow the big squares to randomly generate them.

Pros: Random artwork based on Piet Mondrian's artistic expression and color palette, in accordance with his operating principles.

Cons: Due to the random generation in the form of Procedural Generation, it is not possible to make the final artwork 1:1 with the original artwork by generatingBigBlocks.