Individual Iteration 2
Rectangular blocks are drawn by for loops to randomly select positions and colors in a predefined grid system as a way of generating random colored small and large squares, after the generation of random white lines, random squares are used to fill the canvas, using the square squares formed by the primary colors of red, yellow and blue in Piet Mondrian's work, while he also used black, white and gray to emphasize the purity of the picture's colors. In this way, he creates and recovers his technique, a painting formula consisting of horizontal lines, vertical lines, three colors and three non-colors, known as the grid structure.